change indexer to size_t type

+++ src/app/clusters/time-synchronization-server/time-synchronization-server.cpp
@@ -408,7 +408,7 @@ CHIP_ERROR TimeSynchronizationServer::SetDSTOffset(const DataModel::DecodableLis
 
     auto newDstL   = dstL.begin();
     auto & dstList = mDstOffsetObj.dstOffsetList;
-    uint8_t i      = 0;
+    size_t i       = 0;
 
     while (newDstL.Next())
     {
